[Refactor] 컴포넌트 구조 수정 1. 중복해서 사용하고 있는 컴포넌트만 추출 2. 이메일, 소셜 로그인 화면 각각의 컴포넌트 패키지를 상위(login)의 컴포넌트 패키지로 통합 3. 유틸 패키지 추가 4. 헷갈릴 수 있는 이름들 수정
