+++ app/src/main/java/kr/co/lion/modigm/ui/login/email/EmailLoginFragment.kt
@@ -11,6 +11,7 @@ import android.view.inputmethod.EditorInfo
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
@@ -46,24 +47,33 @@ class EmailLoginFragment : VBBaseFragment<FragmentEmailLoginBinding>(FragmentEma
                 val emailLoginResult by viewModel.emailLoginResult.observeAsState(false)
                 val emailLoginError by viewModel.emailLoginError.observeAsState()
 
+
+                LaunchedEffect(emailLoginResult) {
+                    if (emailLoginResult) {
+                        navigateToBottomNaviFragment(JoinType.EMAIL)
+                    }
+                }
+
+                LaunchedEffect(emailLoginError) {
+                    emailLoginError?.let { error ->
+                        showLoginErrorDialog(error.message.toString())
+                    }
+                }
+
                 Scaffold(
                     modifier = Modifier.fillMaxSize(),
                 ) { innerPadding ->
                     EmailLoginScreen(
                         modifier = Modifier.padding(innerPadding),
                         isLoading = isLoading,
-                        emailLoginResult = emailLoginResult,
-                        emailLoginError = emailLoginError,
-                        onNavigateToBottomNaviFragment = { joinType -> navigateToBottomNaviFragment(joinType) },
-                        onNavigateToFindEmailFragment = { navigateToFindEmailFragment() },
-                        onNavigateToFindPasswordFragment = { navigateToFindPasswordFragment() },
-                        onNavigateToJoinFragment = { joinType -> navigateToJoinFragment(joinType) },
+                        onFindEmailButtonClick = { navigateToFindEmailFragment() },
+                        onFindPasswordButtonClick = { navigateToFindPasswordFragment() },
                         onEmailLoginButtonClick = { email, password, autoLoginValue ->
                             requireActivity().hideSoftInput()
                             viewModel.emailLogin(email, password, autoLoginValue)
                         },
-                        onNavigateToSocialLoginFragment = { navigateToSocialLoginFragment() },
-                        showLoginErrorDialog = { error -> showErrorDialog(error) },
+                        onBackButtonClick = { navigateToSocialLoginFragment() },
+                        onJoinButtonClick = { joinType -> navigateToJoinFragment(joinType) },
                     )
                 }
             }
