--- app/src/main/java/kr/co/lion/modigm/ui/login/social/component/GithubLoginButton.kt
@@ -1,64 +1,0 @@
-package kr.co.lion.modigm.ui.login.social.component
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import kr.co.lion.modigm.R
-import kr.co.lion.modigm.ui.login.social.dpToSp
-
-@Composable
-fun GithubLoginButton(onClick: () -> Unit) {
-    Button(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 30.dp)
-            .height(48.dp),
-        onClick = onClick,
-        colors = ButtonDefaults.buttonColors(
-            containerColor = Color.Black,
-            contentColor = Color.White
-        ),
-        shape = RoundedCornerShape(8.dp),
-        contentPadding = PaddingValues(0.dp),
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(horizontal = 16.dp)
-        ) {
-            Icon(
-                modifier = Modifier
-                    .size(24.dp)
-                    .align(Alignment.CenterStart),
-                painter = painterResource(id = R.drawable.icon_github_logo),
-                contentDescription = "깃허브 로고",
-                tint = Color.White
-            )
-            Text(
-                modifier = Modifier
-                    .align(Alignment.Center)
-                    .offset(x = 12.dp),
-                text = "깃허브 로그인",
-                fontSize = dpToSp(20.dp),
-                color = Color.White,
-
-            )
-        }
-    }
-}
