--- app/src/main/java/kr/co/lion/modigm/ui/login/social/component/KakaoLoginButton.kt
@@ -1,42 +1,0 @@
-package kr.co.lion.modigm.ui.login.social.component
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import kr.co.lion.modigm.R
-
-@Composable
-fun KakaoLoginButton(
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit
-) {
-    Button(
-        onClick = onClick,
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(top = 40.dp),
-        colors = ButtonDefaults.buttonColors(containerColor = Color.Transparent),
-        contentPadding = PaddingValues(0.dp),
-        shape = RoundedCornerShape(0.dp),
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.kakao_login_large_wide),
-            contentDescription = "카카오 로그인",
-            modifier = modifier
-                .fillMaxWidth()
-                .height(48.dp),
-            contentScale = ContentScale.FillBounds
-        )
-    }
-}
