--- app/src/main/java/kr/co/lion/modigm/ui/login/social/component/NavigateToEmailLoginButton.kt
@@ -1,52 +1,0 @@
-package kr.co.lion.modigm.ui.login.social.component
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.layout.wrapContentWidth
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import kr.co.lion.modigm.R
-import kr.co.lion.modigm.ui.login.social.dpToSp
-
-@Composable
-fun NavigateToEmailLoginButton(
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit) {
-    Column(
-        modifier = modifier
-            .fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        TextButton(
-            onClick = onClick,
-            modifier = modifier
-                .wrapContentWidth()
-                .wrapContentHeight()
-                .padding(top = 20.dp),
-            colors = ButtonDefaults.textButtonColors(
-                contentColor = Color.White
-            ),
-            shape = RoundedCornerShape(0.dp),
-        ) {
-            Text(
-                text = "다른 방법으로 로그인",
-                fontSize = dpToSp(16.dp),
-                fontFamily = FontFamily(Font(R.font.one_mobile_pop_otf)),
-                fontWeight = FontWeight.Normal,
-                color = Color.White
-            )
-        }
-    }
-}
