--- app/src/main/java/kr/co/lion/modigm/ui/login/social/component/SocialLoginLoading.kt
@@ -1,29 +1,0 @@
-package kr.co.lion.modigm.ui.login.social.component
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.input.pointer.pointerInput
-
-@Composable
-fun SocialLoginLoading(
-    modifier: Modifier = Modifier,
-    isLoading: Boolean
-) {
-    if (isLoading) {
-        Box(
-            modifier = modifier
-                .fillMaxSize()
-                .background(Color.Black.copy(alpha = 0.5f))
-                .pointerInput(Unit) { },
-            contentAlignment = Alignment.Center
-        ) {
-            CircularProgressIndicator(color = Color.White)
-        }
-    }
-}
